watson-service/requirements.txt
appdirs==1.4.4
certifi==2020.4.5.1
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
distlib==0.3.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
ibm-cloud-sdk-core==1.5.1
ibm-watson==4.5.0
identify==1.4.16
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
nodeenv==1.3.5
Pillow==7.2.0
pre-commit==2.4.0
PyJWT==1.7.1
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
toml==0.10.1
urllib3==1.25.9
virtualenv==20.0.21
websocket-client==0.48.0
Werkzeug==1.0.1
